\name{about}
\title{About GCDkit}
\alias{about}

\description{
Prints short information about the current version of GCDkit and contact
addresses of its authors.
}

\usage{
about()
}

\arguments{
None.
}

\concept{Menu: GCDkit}

\keyword{misc}

\author{Vojtech Janousek, \email{vojtech.janousek@geology.cz}}
