\name{appendSingle}
\title{Append empty label or variable}
\alias{appendSingle}

\description{
Appends an empty numeric data column or a new label to the current data set.
}

\usage{
appendSingle()
}

\value{
Returns the corrected version of the data frame '\code{labels}' or numeric matrix '\code{WR}'.
}

\keyword{manip}
\concept{Edit}
\concept{Menu: Data handling}
\author{Vojtech Janousek, \email{vojtech.janousek@geology.cz}}
