\title{List of available classification schemes}
\name{.claslist}
\alias{.claslist}

\description{
The function returns a list of  classification diagrams available in the system.
}

\usage{
.claslist()
}

\value{
A matrix with two columns:
  \item{menu}{menu items}
  \item{function}{the attached functions}
}

\author{Vojtech Erban, \email{vojtech.erban@geology.cz}}

\keyword{manip}
\concept{Grouping}
\concept{Classification}
\concept{Plot}
\concept{Figaro}
\concept{Subsetting}
