\name{Edit labels}
\title{Edit labels}
\alias{editLabels}

\description{
    Simultaneous editing of all labels using a spreadsheet-like interface.
}

\usage{
    editLabels()
}

\arguments{
  \item{}{none.}
}

\details{ 
    The function invokes a spreadsheet-like interface that enables the user to edit 
    the labels for individual samples. When all the desired changes have been 
    performed, close button is to be clicked.
}

\value{
    Returns the corrected version of the data frame '\code{labels}'.
}

\seealso{
    '\code{\link{data.entry}}'
}

\keyword{manip}
\concept{Edit}
\concept{Menu: Data handling}
\author{Vojtech Janousek, \email{vojtech.janousek@geology.cz}}
