\name{quitGCDkit}
\title{Exit GCDkit}
\alias{quitGCDkit}

\description{
Exits GCDkit (nicely).
}

\usage{
quitGCDkit()
}

\arguments{
None.
}

\section{Menu}{
GCDkit: Exit GCDkit
}

\details{
By invoking this command the user is not prompted whether he wants to save his unfinished 
work in the 'Workspace image', i.e. file '\code{.RData}' in the main GCDkit directory.
}

\seealso{
'\code{\link{quit}}'
}

\keyword{misc}
