\name{figaroMenu}
\alias{figaroOn}
\alias{figaroOff}
\title{Controlling the menu Plot editing}

\description{
Functions controlling the activity that render some options of the menu
'\code{Plot editing}' on or off, as appropriate.
}


\usage{
figaroOn(keep.plate=FALSE)

figaroOff()
}

\arguments{
\item{keep.plate}{logical; should be any existing plate definition kept?}
}

\value{
None.
}

\seealso{
    '\code{\link{figaro}}'
}

\keyword{internal}
\keyword{iplot}
\keyword{aplot}
\concept{Plot}
\concept{Figaro}
\concept{Menu: Plot editing}

\author{
Colin M. Farrow, \email{colinfarrow537@gmail.com} 

and Vojtech Janousek, \email{vojtech.janousek@geology.cz}
}
