\name{figSave}
\title{Saving a Figaro plot}
\alias{figSave}

\description{
Saves the current Figaro-compatible plot, both the template and the data needed 
for the plotting ('\code{x.data}', '\code{y.data}'). 
}

\usage{
figSave()
}

\arguments{
None.
}

\details{
The default suffix for the saved diagrams is '\code{fgr}'.
}

\seealso{
  \link{figLoad}
  \link{figaro}
}

\keyword{hplot}
\keyword{file}
\concept{Plot}
\concept{Figaro}
\concept{File}
\concept{Menu: Plot editing}

\author{
Colin M. Farrow, \email{colinfarrow537@gmail.com} 

and Vojtech Janousek, \email{vojtech.janousek@geology.cz}
}
