\name{Geometric mean}
\alias{gMean}
\title{Geometric mean}

\description{
    A function to calculate the geometric mean.   
}

\usage{
gMean(x)
}

\arguments{
  \item{x}{a numeric vector}
}

\details{
    The function returns a geometric mean of the given numeric vector. The \code{NA} values are removed beforehand. 
}

\value{
    Geometric mean. 
}

\seealso{
    \code{\link{mean}}
}


\author{Vojtech Janousek, \email{vojtech.janousek@geology.cz}}

\keyword{internal}
\keyword{univar}
\concept{Menu: Calculations: Statistics}
\concept{Statistics}
