\name{sazava}
\docType{data}
\alias{sazava}
\title{Whole-rock composition of the Sazava suite, Central Bohemian Plutonic Complex}
\description{
    This data set gives the whole-rock major- and trace-element contents in selected
    samples (gabbros, quartz diorites, tonalites and trondhjemites) of the  c. 355 My 
    old calc-alkaline Sazava suite of the Variscan Central Bohemian Plutonic Complex 
    (Bohemian Massif, Czech Republic). 
}
\usage{
data(sazava)
}

\format{A data frame containing 14 observations.}

\examples{
data(sazava)
accessVar("sazava")
binary("SiO2","Ba")
}


\source{Vojtech Janousek, \email{vojtech.janousek@geology.cz}}

\references{
Janousek V, Rogers G,  Bowes DR (1995) Sr-Nd isotopic constraints on the petrogenesis of the Central Bohemian Pluton, 
Czech Republic. Geol Rundsch 84: 520-534
\href{http://dx.doi.org/10.1007/BF00284518}{doi: 10.1007/BF00284518} 


Janousek V, Bowes DR, Rogers G, Farrow CM, Jelinek E (2000) Modelling diverse processes in the petrogenesis 
of a composite batholith: the Central Bohemian Pluton, Central European Hercynides. J Petrol 41: 511-543
\href{http://dx.doi.org/10.1093/petrology/41.4.511}{doi: 10.1093/petrology/41.4.511} 

Janousek V, Braithwaite CJR, Bowes DR, Gerdes A (2004) Magma-mixing in the genesis of Hercynian calc-alkaline 
granitoids: an integrated petrographic and geochemical study of the Sazava intrusion, Central Bohemian Pluton, 
Czech Republic. Lithos 78: 67-99
\href{http://dx.doi.org/10.1016/j.lithos.2004.04.046}{doi: 10.1016/j.lithos.2004.04.046} 
}
\keyword{datasets}
